Various fixes and changes -> more stable results
We now support multiple matrix trials;

We now share one sample feature matrix etc. across all conditional
models in a density family;

Resampling is now done only once at the start of the fitting process
if matrixtrials == 1, as expected.

The fitting process seems to work more reliably now instead of sometimes
having the objective diverge to -inf. We could also try changing the
objective function to the grad norm if divergence to -inf continues to
be an issue.